lab4/app/requirements.txt
flask==2.3.3
waitress
flask-sqlalchemy
mysqlclient
sqlalchemy-utils
jinja2
jsonify
flask-restx
